requirements.txt
pyjwt
requests
msal
python-dotenv
flask
pyodbc
pymilvus
neo4j
flasgger
transformers
torch
librosa
Pillow
pypdf
numpy
PyPDF2
pytesseract
SpeechRecognition
marshmallow
sentence-transformers
tf-keras
black
isort
ruff
opencensus-ext-azure
opencensus-ext-flask
azure-keyvault-secrets
aioodbc
gunicorn
pytest-asyncio